+++ src/index.ts
@@ -1,2 +1,3 @@
+export * from "./history";
 export * from "./queue";
 export * from "./ringBuffer";